requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
django-rq==2.4.0
gunicorn==20.0.4
channels-redis==3.2.0
rq~=1.7.0
channels~=3.0.3
opencv-python ==4.5.2.54
numpy == 1.19.5
PyQt5==5.15.1
PyQt5-sip==12.8.1
qimage2ndarray~=1.8.3
matplotlib == 3.4.2